requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
spotipy>=2.23.0
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
Pillow>=10.0.0
